ratings work, still need to edit user (admin)

+++ controllers/UserController.js
@@ -63,17 +63,18 @@ module.exports = {
 
   getList: function (request, response) {
     try {
-     
-      const { limit, page } = request.body
+      const page = request.query.page || 1
+      const limit = request.query.limit || 12
+
       const pageCheck = page > 0 ? page-1 : 0 
       const offset = limit * pageCheck 
 
       User.findAndCountAll({
         limit,
         offset
-      }).then(users => {
+      }).then(data => {
         response.status(200).send({
-          data: users, 
+          data: data, 
           message: "Users found"
         })
       })
@@ -154,12 +155,24 @@ module.exports = {
   },
   removeSingle: function (request, response) {
     try {
-      const { username } = request.params;
+
+      // non-admins are not allowed to delete a user
       if (!request.user.admin) {
         response.status(401).send({
           error: "Permission denied",
         });
         return;
+      }
+
+      const { username } = request.params;
+      const user = request.user;
+
+      // a user should not be allowed to delete themself
+      if (user.username === username) {
+        response.status(401).send({
+          error: "You cannot delete yourself",
+        });
+        return
       }
 
       User.destroy({
@@ -175,6 +188,12 @@ module.exports = {
         }
         response.status(200).send({
           message: "User removed",
+        });
+      })
+      .catch(error => {
+        console.log("Error deleting user", error);
+        response.status(500).send({
+          error: 'Error deleting User'
         });
       });
     } catch (error) {
